+++ R/object_usage_linter.R
@@ -85,8 +85,16 @@ object_usage_linter <- function(interpret_glue = NULL, interpret_extensions = c(
 
     xml <- source_expression$full_xml_parsed_content
 
+    # Catch missing packages and report them as lints
+    library_lints <- list()
+    lint_hook <- function(lint_node, lint_msg) {
+      library_lints[[length(library_lints) + 1L]] <<- xml_nodes_to_lints(
+        lint_node, source_expression = source_expression, lint_message = lint_msg, type = "warning"
+      )
+    }
+
     # run the following at run-time, not "compile" time to allow package structure to change
-    env <- make_check_env(pkg_name, xml)
+    env <- make_check_env(pkg_name, xml, lint_hook)
 
     fun_assignments <- xml_find_all(xml, xpath_function_assignment)
 
@@ -139,17 +147,20 @@ object_usage_linter <- function(interpret_glue = NULL, interpret_extensions = c(
       nodes[missing_symbol] <- lapply(which(missing_symbol), function(i) {
         line_based_match <- xml_find_first(
           fun_assignment,
-          glue::glue_data(res[i, ], "descendant::expr[@line1 = {line1} and @line2 = {line2}]")
+          glue::glue_data(res[i,], "descendant::expr[@line1 = {line1} and @line2 = {line2}]")
         )
         if (is.na(line_based_match)) fun_assignment else line_based_match
       })
 
-      xml_nodes_to_lints(nodes, source_expression = source_expression, lint_message = res$message, type = "warning")
+      c(
+        library_lints,
+        xml_nodes_to_lints(nodes, source_expression = source_expression, lint_message = res$message, type = "warning")
+      )
     })
   })
 }
 
-make_check_env <- function(pkg_name, xml) {
+make_check_env <- function(pkg_name, xml, lint_hook) {
   if (!is.null(pkg_name)) {
     parent_env <- try_silently(getNamespace(pkg_name))
   }
@@ -160,7 +171,7 @@ make_check_env <- function(pkg_name, xml) {
 
   symbols <- c(
     get_assignment_symbols(xml),
-    get_imported_symbols(xml)
+    get_imported_symbols(xml, lint_hook)
   )
 
   # Just assign them an empty function
@@ -245,7 +256,7 @@ parse_check_usage <- function(expression,
     ))
   }
   # nocov end
-  res <- res[!is_missing, ]
+  res <- res[!is_missing,]
 
   res$line1 <- ifelse(
     nzchar(res$line1),
@@ -264,7 +275,7 @@ parse_check_usage <- function(expression,
   res
 }
 
-get_imported_symbols <- function(xml) {
+get_imported_symbols <- function(xml, lint_hook) {
   import_exprs_xpath <- "
   //SYMBOL_FUNCTION_CALL[text() = 'library' or text() = 'require']
     /parent::expr
@@ -280,10 +291,17 @@ get_imported_symbols <- function(xml) {
   import_exprs <- xml_find_all(xml, import_exprs_xpath)
   imported_pkgs <- get_r_string(import_exprs)
 
-  unlist(lapply(imported_pkgs, function(pkg) {
+  unlist(lapply(seq_along(import_exprs), function(i) {
     tryCatch(
-      getNamespaceExports(pkg),
-      error = function(e) character()
+      getNamespaceExports(imported_pkgs[[i]]),
+      error = function(e) {
+        pkg <- imported_pkgs[[i]]
+        lint_node <- xml_parent(import_exprs[[i]])
+        lint_msg <- paste0("Could not find exported symbols for package \"", pkg, "\" in library ",
+                           toString(shQuote(.libPaths())), ". This may lead to false positives.")
+        lint_hook(lint_node, lint_msg)
+        character()
+      }
     )
   }))
 }