+++ tests/testthat/test-object_usage_linter.R
@@ -254,7 +254,6 @@ test_that("used symbols are detected correctly", {
   )
 
 
-
   # regression #1322
   expect_silent(expect_lint("assign('x', 42)", NULL, object_usage_linter()))
 })
@@ -648,7 +647,10 @@ test_that("missing libraries don't cause issue", {
         a
       }
     "),
-    NULL,
+    list(
+      "Could not find exported symbols for package \"a.a.a.z.z.z\"",
+      line_number = 1L
+    ),
     object_usage_linter()
   )
 })
@@ -885,15 +887,15 @@ test_that("dplyr's .env-specified objects are marked as 'used'", {
 
 test_that("interpret_glue is deprecated", {
   expect_warning(
-    {
-      linter_no <- object_usage_linter(interpret_glue = FALSE)
-    },
+  {
+    linter_no <- object_usage_linter(interpret_glue = FALSE)
+  },
     rex::rex("interpret_glue", anything, "deprecated")
   )
   expect_warning(
-    {
-      linter_yes <- object_usage_linter(interpret_glue = TRUE)
-    },
+  {
+    linter_yes <- object_usage_linter(interpret_glue = TRUE)
+  },
     rex::rex("interpret_glue", anything, "deprecated")
   )